Add files via upload
This study extends the analysis of Altindag, Erten, and Keskin (2022), focusing on the mental health impacts of a stringent age-specific lockdown on adults aged 65 and over in Turkey. Utilizing a regression discontinuity framework, we replicate their finding of a significant negative impact on mental health, with a decline in mobility leading to an increase in both somatic and nonsomatic symptoms of mental distress by roughly 0.2 standard deviations. Enhancing this approach, our analysis introduces an interaction term to examine the differential effects on married individuals. Additionally, we construct a religiosity index to assess whether religious beliefs buffer the adverse effects on mental well-being during the lockdown. These modifications aim to deepen the understanding of the complex interplay between personal characteristics, mental health, and mobility restrictions in the context of the COVID-19 pandemic.